♻️ [refactor]#382: 마이페이지 소개 컴포넌트 관심사에 맞게 합침
