+++ client/src/pages/ProfilePage/index.tsx
@@ -2,8 +2,6 @@ import React from "react";
 import { useParams } from "react-router-dom";
 import PageTemplate from "@pages/PageTemplate";
 import MyPageUserProfile from "@components/MyPageUserProfile";
-import MyPageUserIntroduce from "@components/MyPageUserIntroduce";
-import MyPageSubInfoContainer from "@components/MyPageSubInfoContainer";
 import MyPageEditButton from "@components/MyPageEditButton";
 import useUserInfo from "@hooks/query/useUserInfo";
 import FollowButton from "@components/FollowButton";
@@ -21,8 +19,6 @@ const ProfilePage = () => {
     <PageTemplate isRoot={isOwner}>
       <s.MyProfileContainer>
         <MyPageUserProfile userId={profileUserId} isOwner={isOwner} />
-        <MyPageUserIntroduce userId={profileUserId} isOwner={isOwner} />
-        <MyPageSubInfoContainer userId={profileUserId} />
         <s.ButtonContainer>
           {isOwner ? (
             <MyPageEditButton userId={profileUserId} ownerId={id} isOwner={isOwner} />
